+++ HodEngine/src/Scene.cpp
@@ -71,16 +71,16 @@ void Scene::drawDebugPhysics(CameraComponent* camera, float dt)
             line.vertices[0].pos[0] = pxLine.pos0.x;
             line.vertices[0].pos[1] = pxLine.pos0.y;
             line.vertices[0].pos[2] = pxLine.pos0.z;
-            line.vertices[0].color[0] = (pxLine.color0 & 0x00FF0000) >> 16;
-            line.vertices[0].color[1] = (pxLine.color0 & 0x0000FF00) >> 8;
-            line.vertices[0].color[2] = (pxLine.color0 & 0x000000FF) >> 0;
+            line.vertices[0].color[0] = (float)((pxLine.color0 & 0x00FF0000) >> 16);
+            line.vertices[0].color[1] = (float)((pxLine.color0 & 0x0000FF00) >> 8);
+            line.vertices[0].color[2] = (float)((pxLine.color0 & 0x000000FF) >> 0);
 
             line.vertices[1].pos[0] = pxLine.pos1.x;
             line.vertices[1].pos[1] = pxLine.pos1.y;
             line.vertices[1].pos[2] = pxLine.pos1.z;
-            line.vertices[1].color[0] = (pxLine.color1 & 0x00FF0000) >> 16;
-            line.vertices[1].color[1] = (pxLine.color1 & 0x0000FF00) >> 8;
-            line.vertices[1].color[2] = (pxLine.color1 & 0x000000FF) >> 0;
+            line.vertices[1].color[0] = (float)((pxLine.color1 & 0x00FF0000) >> 16);
+            line.vertices[1].color[1] = (float)((pxLine.color1 & 0x0000FF00) >> 8);
+            line.vertices[1].color[2] = (float)((pxLine.color1 & 0x000000FF) >> 0);
         }
 
         this->drawLine(lines, 0.0f);
@@ -103,23 +103,23 @@ void Scene::drawDebugPhysics(CameraComponent* camera, float dt)
             tri.vertices[0].pos[0] = pxTri.pos0.x;
             tri.vertices[0].pos[1] = pxTri.pos0.y;
             tri.vertices[0].pos[2] = pxTri.pos0.z;
-            tri.vertices[0].color[0] = (pxTri.color0 & 0x00FF0000) >> 16;
-            tri.vertices[0].color[1] = (pxTri.color0 & 0x0000FF00) >> 8;
-            tri.vertices[0].color[2] = (pxTri.color0 & 0x000000FF) >> 0;
+            tri.vertices[0].color[0] = (float)((pxTri.color0 & 0x00FF0000) >> 16);
+            tri.vertices[0].color[1] = (float)((pxTri.color0 & 0x0000FF00) >> 8);
+            tri.vertices[0].color[2] = (float)((pxTri.color0 & 0x000000FF) >> 0);
 
             tri.vertices[1].pos[0] = pxTri.pos1.x;
             tri.vertices[1].pos[1] = pxTri.pos1.y;
             tri.vertices[1].pos[2] = pxTri.pos1.z;
-            tri.vertices[1].color[0] = (pxTri.color1 & 0x00FF0000) >> 16;
-            tri.vertices[1].color[1] = (pxTri.color1 & 0x0000FF00) >> 8;
-            tri.vertices[1].color[2] = (pxTri.color1 & 0x000000FF) >> 0;
+            tri.vertices[1].color[0] = (float)((pxTri.color1 & 0x00FF0000) >> 16);
+            tri.vertices[1].color[1] = (float)((pxTri.color1 & 0x0000FF00) >> 8);
+            tri.vertices[1].color[2] = (float)((pxTri.color1 & 0x000000FF) >> 0);
 
             tri.vertices[2].pos[0] = pxTri.pos2.x;
             tri.vertices[2].pos[1] = pxTri.pos2.y;
             tri.vertices[2].pos[2] = pxTri.pos2.z;
-            tri.vertices[2].color[0] = (pxTri.color2 & 0x00FF0000) >> 16;
-            tri.vertices[2].color[1] = (pxTri.color2 & 0x0000FF00) >> 8;
-            tri.vertices[2].color[2] = (pxTri.color2 & 0x000000FF) >> 0;
+            tri.vertices[2].color[0] = (float)((pxTri.color2 & 0x00FF0000) >> 16);
+            tri.vertices[2].color[1] = (float)((pxTri.color2 & 0x0000FF00) >> 8);
+            tri.vertices[2].color[2] = (float)((pxTri.color2 & 0x000000FF) >> 0);
         }
 
         this->drawTri(tris, 0.0f);
